requirements.txt
Flask==1.0.2
facebookads==2.11.4
python-dotenv==0.8.2
ccxt==